# openapi/swagger.yaml
basePath: /v1
definitions:
...
  plan.CalendarResponse:
    properties:
      days:
        items:
          $ref: '#/definitions/plan.CalendarDay'
        type: array
      end:
        type: string
      start:
        type: string
    type: object
...
  plan.CreateScheduleItemInput:
    properties:
      category:
        $ref: '#/definitions/plan.ScheduleCategory'
      color:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      end_time:
        type: string
      notes:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      student_id:
        type: string
      subject_id:
        type: string
      title:
        maxLength: 200
        type: string
    required:
    - start_date
    - title
    type: object
...
  plan.DayViewResponse:
    properties:
      activities:
        items:
          $ref: '#/definitions/plan.ActivitySummary'
        type: array
      attendance:
        $ref: '#/definitions/plan.AttendanceSummary'
      date:
        type: string
      events:
        items:
          $ref: '#/definitions/plan.EventSummary'
        type: array
      schedule_items:
        items:
          $ref: '#/definitions/plan.ScheduleItemResponse'
        type: array
    type: object
...
  plan.ScheduleItemListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/plan.ScheduleItemResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
    type: object
  plan.ScheduleItemResponse:
    properties:
      category:
        $ref: '#/definitions/plan.ScheduleCategory'
      color:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      end_time:
        type: string
      id:
        type: string
      is_completed:
        type: boolean
      linked_activity_id:
        type: string
      notes:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      student_id:
        type: string
      student_name:
        type: string
      subject_id:
        type: string
      title:
        type: string
    type: object
...
  plan.WeekViewResponse:
    properties:
      days:
        items:
          $ref: '#/definitions/plan.CalendarDay'
        type: array
      week_end:
        type: string
      week_start:
        type: string
    type: object
...
  shared.AppError:
    properties:
      code:
        type: string
      err:
        description: wrapped internal error, never exposed to client
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
host: localhost:3500
info:
  contact: {}
  description: API for the Homegrown Academy homeschooling platform
  title: Homegrown Academy API
  version: 0.1.0
paths:
...
  /planning/calendar:
    get:
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        type: string
      - description: Filter by student ID
        in: query
        name: student_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plan.CalendarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Get aggregated calendar
      tags:
      - planning
  /planning/calendar/day/{date}:
    get:
      parameters:
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      - description: Filter by student ID
        in: query
        name: student_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plan.DayViewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Get day view
      tags:
      - planning
  /planning/calendar/pdf:
    get:
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      - description: Filter by student ID
        in: query
        name: student_id
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: PDF file
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Export calendar as PDF
      tags:
      - planning
  /planning/calendar/print:
    get:
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      - description: Filter by student ID
        in: query
        name: student_id
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Get print-friendly calendar
      tags:
      - planning
  /planning/calendar/week/{date}:
    get:
      parameters:
      - description: Any date within the week (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      - description: Filter by student ID
        in: query
        name: student_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plan.WeekViewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Get week view
      tags:
      - planning
  /planning/schedule-items:
    get:
      parameters:
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Filter by student ID
        in: query
        name: student_id
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by completion status
        in: query
        name: is_completed
        type: boolean
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Results per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plan.ScheduleItemListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: List schedule items
      tags:
      - planning
    post:
      consumes:
      - application/json
      parameters:
      - description: Schedule item details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/plan.CreateScheduleItemInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Create a schedule item
      tags:
      - planning
  /planning/schedule-items/{id}:
    delete:
      parameters:
      - description: Schedule item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Delete a schedule item
      tags:
      - planning
    get:
      parameters:
      - description: Schedule item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plan.ScheduleItemResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Get a schedule item
      tags:
      - planning
    patch:
      consumes:
      - application/json
      parameters:
      - description: Schedule item ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/plan.UpdateScheduleItemInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Update a schedule item
      tags:
      - planning
  /planning/schedule-items/{id}/complete:
    patch:
      parameters:
      - description: Schedule item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/shared.AppError'
      security:
      - BearerAuth: []
      summary: Mark a schedule item as completed
      tags:
      - planning
...
swagger: "2.0"